Add setting the Sonarr IP/Hostname from a file.
Added a setting to read your Sonarr host's IP or hostname from a file. Why? Because if you run Sonarr on a host/container with a dynamic IP address this allows the script to find it. You can have a small script on Sonarr's host to place a text file with its current IP at a location this script can find it. If not, just make a text file with the IP/hostname along side this script and you're ready to go.